requirements.txt
aiomysql==0.1.1
aiosqlite==0.17.0
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.2
black==22.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==37.0.4
dnspython==2.2.1
docstring-parser==0.12
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.78.0
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mega.py==1.0.8
mypy-extensions==0.4.3
ntc-templates==3.0.0
orjson==3.6.7
paramiko==2.11.0
passlib==1.7.4
pathlib==1.0.1
pathspec==0.9.0
pip-autoremove==0.10.0
platformdirs==2.5.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pymongo==3.12.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pypika-tortoise==0.1.5
pyserial==3.5
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.1
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
scp==0.14.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
starlette==0.19.1
targ==0.3.7
tenacity==5.1.5
text-unidecode==1.3
textfsm==1.1.2
tomli==2.0.1
typesystem==0.4.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
Werkzeug==2.1.2
zipp==3.7.0
